% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{antarctic_coastline}
\alias{antarctic_coastline}
\title{Antarctic Coastline Data}
\format{
An \code{sf} object with X rows and Y columns (adjust to your dataset)
}
\usage{
data(antarctic_coastline)
}
\description{
This dataset contains the Antarctic coastline as a spatial object.
}
\examples{
\dontrun{
data(antarctic_coastline)
plot(antarctic_coastline)
}
}
\keyword{datasets}
