% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_ice_shelves.R
\docType{data}
\name{antarctic_ice_shelves}
\alias{antarctic_ice_shelves}
\title{Antarctic Ice Shelves Data}
\format{
An \code{sf} object with X rows and Y columns (adjust to your dataset)
}
\usage{
data(antarctic_ice_shelves)
}
\description{
This dataset contains the Antarctic Ice Shelves as a spatial object.
}
\examples{
\dontrun{
data(antarctic_ice_shelves)
plot(antarctic_ice_shelves)
}
}
\keyword{datasets}
