% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarcticR.R
\name{clusterResult}
\alias{clusterResult}
\title{A function to use some clustering methods from the dbscan package}
\usage{
clusterResult(haversineMatrix, eps = 2e+05, minPts, eps_cl)
}
\description{
A function to use some clustering methods from the dbscan package
}
\examples{

}
\keyword{Antarctica,}
\keyword{draw,}
\keyword{plot}
