% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarcticR.R
\name{csvToDF}
\alias{csvToDF}
\title{Turn a longitude, latitude csv file into a dataframe}
\usage{
csvToDF(csvFile)
}
\arguments{
\item{csvFile}{Your csv file}
}
\value{
A dataframe
}
\description{
Generate a dataframe from a longitude-latitude csv file
}
\examples{
df <- csvToHaversineMat("myData.csv")
}
\keyword{dataframe}
