% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarcticR.R
\name{csvToHaversineMat}
\alias{csvToHaversineMat}
\title{A function to generate a Haversine matrix from a csv file}
\usage{
csvToHaversineMat(csvFile)
}
\arguments{
\item{df}{Your csv file}
}
\value{
A haversine distance matrix
}
\description{
Generate a distance matrix of great-circle distances from a csv file with longitude and latitude distances
}
\examples{

}
\keyword{Haversine,}
\keyword{csv,}
\keyword{distance}
\keyword{matrix,}
