% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarcticR.R
\name{drawAntarctica}
\alias{drawAntarctica}
\title{Set up the drawing of a map of Antarctica}
\usage{
drawAntarctica()
}
\description{
Set up the drawing of a map of Antarctica
}
\examples{
world3 <- drawAntarctica()
plot(world3)
}
\keyword{Antarctica,}
\keyword{draw,}
\keyword{plot}
