% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarcticR.R
\name{drawBedmap}
\alias{drawBedmap}
\title{Set up the drawing of a map of Antarctica BEDMAP2}
\usage{
drawBedmap(BEDMAP_GRAD = "thickness", reduceResolutionBy = 5)
}
\description{
Set up the drawing of a map of Antarctica BEDMAP2
}
\examples{
world3 <- drawAntarctica()
world3
}
\keyword{Antarctica,}
\keyword{draw,}
\keyword{plot}
