% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarcticR.R
\name{genCartesianMat}
\alias{genCartesianMat}
\title{A function to generate a Cartesian matrix from a dataframe}
\usage{
genCartesianMat(df)
}
\arguments{
\item{df}{Your data frame}
}
\value{
A Cartesian distance matrix
}
\description{
Generate a distance matrix of x-y-z distances from a dataframe with longitude and latitude points
}
\examples{

points <- read.csv("dividedEvents1.csv",header=T, sep=",")
df.points <- as.matrix(points)
antFrame = data.frame(df.points)
print("Computing distance matrix...")
d  <- genCartesianMatrix(antFrame)

}
\keyword{Cartesian,}
\keyword{dataframe,}
\keyword{distance}
\keyword{matrix,}
