% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarcticR.R
\name{genHaversineMat}
\alias{genHaversineMat}
\title{A function to generate a Haversine matrix from a dataframe}
\usage{
genHaversineMat(df)
}
\arguments{
\item{df}{Your data frame}
}
\value{
A haversine distance matrix
}
\description{
Generate a distance matrix of great-circle distances from a dataframe with longitude and latitude distances
}
\examples{

points <- read.csv("dividedEvents1.csv",header=T, sep=",")
df.points <- as.matrix(points)
antFrame = data.frame(df.points)
print("Computing distance matrix...")
require(geosphere)
d  <- genHaversineMat(antFrame)

}
\keyword{Haversine,}
\keyword{dataframe,}
\keyword{distance}
\keyword{matrix,}
