% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_antarctic_coastline.R
\name{load_antarctic_coastline}
\alias{load_antarctic_coastline}
\title{Load Antarctic Coastline Data}
\usage{
load_antarctic_coastline()
}
\value{
A spatial or simple features (\code{sf}) object containing the Antarctic coastline.
}
\description{
This function loads the Antarctic coastline dataset.
The data derived from the DLR-German Aerospace Center Icelines Download Service.
}
\examples{
\dontrun{
coastline <- load_antarctic_coastline()
plot(coastline)
}

}
\author{
Sebastian Rothaug
}
