% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_antarctic_continent.R
\name{load_continent}
\alias{load_continent}
\title{Load Antarctica Boundaries}
\usage{
load_continent()
}
\value{
An \code{sf} object containing the geometry of Antarctica.
}
\description{
This function loads Antarctica's country boundaries as a simple features (\code{sf}) object
and transforms the projection to EPSG:3031 (Antarctic Polar Stereographic).
}
\examples{
\dontrun{
continent <- load_continent()
plot(continent)
}


}
\author{
Sebastian Rothaug
}
