% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ice_shelves.R
\name{load_ice_shelves}
\alias{load_ice_shelves}
\title{Load Antarctic Ice Shelf Data}
\usage{
load_ice_shelves()
}
\value{
A spatial or simple features (\code{sf}) object containing all the Antarctic Ice Shelves.
}
\description{
This function loads the Antarctic Ice Shelf dataset.
The data derived from naturalearthdata Download Service, based on the 2003-2004 reference data.
}
\examples{
\dontrun{
ice_shelves <- load_ice_shelves()
plot(ice_shelves)
}

}
\author{
Sebastian Rothaug
}
