% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarcticR.R
\name{longLatToSimpleBEDMAP}
\alias{longLatToSimpleBEDMAP}
\title{A function to convert from lon/lat to the BEDMAP grid}
\usage{
longLatToSimpleBEDMAP(longLatDataFrame)
}
\value{
bedmapFrame
}
\description{
A function to convert from lon/lat to the BEDMAP grid
}
\examples{

}
\keyword{Antarctica,}
\keyword{BEDMAP}
\keyword{draw,}
\keyword{plot,}
