% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_antarctica.R
\name{plot_season}
\alias{plot_season}
\title{Plot Antarctic Sea Ice for a Given Season}
\usage{
plot_season(season)
}
\arguments{
\item{season}{A string indicating the season. Accepts \code{"summer"} or \code{"winter"}.}
}
\value{
A \code{ggplot} object displaying Antarctic sea ice extent for the given season.
}
\description{
This function generates a map of Antarctic sea ice extent for either the summer or winter season.
It visualizes the seasonal minimum (summer) and maximum (winter) ice thickness.
}
\details{
\itemize{
\item Summer ice extent (minimum) is shown in \strong{red}.
\item Winter ice extent (maximum) is shown in \strong{blue}.
\item The Antarctic continent is displayed in \strong{gray}.
}
}
\examples{
\dontrun{
plot_season("summer")
plot_season("winter")
}

}
\seealso{
\code{\link{load_sea_ice}}, \code{\link{load_continent}}
}
\author{
Sebastian Rothaug
}
