% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarcticR.R
\name{drawAntarctica}
\alias{drawAntarctica}
\title{Draw Antarctica}
\usage{
drawAntarctica()
}
\description{
Draw Antarctica
}
\examples{
drawAntarctica
}
\keyword{Antarctica,}
\keyword{draw,}
\keyword{plot}
