% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHaversineMat.R
\name{genHaversineMat}
\alias{genHaversineMat}
\title{A function to generate a Haversine matrix from a dataframe
Generate a distance matrix of great-circle distances from a dataframe with longitude and latitude distances}
\usage{
genHaversineMat(df)
}
\arguments{
\item{df}{Your data frame}
}
\value{
A haversine distance matrix
}
\description{
A function to generate a Haversine matrix from a dataframe
Generate a distance matrix of great-circle distances from a dataframe with longitude and latitude distances
}
\examples{

cities <- read.csv("dividedEvents1.csv",header=T, sep=",")
df.cities <- as.matrix(cities)
antFrame = data.frame(df.cities)#long=df.cities$long, lat=df.cities$lat)
antFrame
print("Computing distance matrix...")
require(geosphere)
d  <- genHaversineMat(antFrame)   # distance matrix

genHaversineMat()
}
\keyword{Haversine,}
\keyword{dataframe,}
\keyword{distance}
\keyword{matrix,}
