% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalF.R
\name{getCalF}
\alias{getCalF}
\title{Download and Plot Antarctic Calving Fronts from DLR Icelines}
\usage{
getCalF(IceShelf, Year = NULL, save_dir = ".")
}
\arguments{
\item{IceShelf}{Character. Name of the Antarctic ice shelf (e.g., "Getz1","GeorgeSouth", "Thwaites1", "Amery").}

\item{Year}{Optional. Numeric or character. Year to download (e.g., 2019) - The annual data availability is not constant and is individual for each ice shelf. If NULL (default), all available years are downloaded.}

\item{save_dir}{Character. Directory where files will be saved. If Year is NULL, a subfolder named after the ice shelf will be created.}
}
\value{
A ggplot object showing the calving fronts for the selected ice shelf and year(s).
}
\description{
Downloads and visualizes Antarctic ice shelf calving front shapefiles from the DLR Icelines project for a specified year or all available years.
}
\details{
When \code{Year} is not specified, the function scrapes the DLR Icelines server to download and plot all available calving front shapefiles for the given ice shelf.
The data are visualized using Antarctic Polar Stereographic projection (EPSG:3031), and the continent is plotted for geographic reference.
If the file(s) already exist locally, they will not be downloaded again.
}
\note{
Requires internet connection and the \code{rvest}, \code{sf}, \code{ggplot2} packages.
}
\examples{
\dontrun{
getCalF("Getz1")              # Download and plot all available years for Getz1 Ice Shelf
getCalF("Wilkins", 2020)      # Download and plot only the year 2020 for Wilkins
}

}
\author{
Sebastian Rothaug
}
