% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarcticR.R
\name{plotAntarctica}
\alias{plotAntarctica}
\title{Plot points on the antarctic map}
\usage{
plotAntarctica(antMap, df, cluster = FALSE)
}
\arguments{
\item{antMap}{your map made from drawAntarctica}

\item{df}{Your data frame}
}
\description{
Plot points on the antarctic map
}
\examples{
world4 <- plotAntarctica(map, dataFrame)
world4
}
\keyword{Antarctica,}
\keyword{draw,}
\keyword{plotx}
